Allows filtering with PostgreSQL network operators
The primary addition of this commit is the new filter operators
corresponding to the PostgreSQL network address operators, but it also
makes several other changes to the testing framework, including changes
to the Travis-CI build configuration.

First, it restructures the unit test base classes a bit. It provides
separate "parallel" base classes for database tests: one for
Flask-SQLAlchemy and one for "pure" SQLAlchemy. This fixes some
confusion in the database configuration for test classes.

Second, it makes some code style changes in the unit tests.

Finally, it forces Travis to install an appropriate version of
PostgreSQL and an appropriate Python library for accessing a PostgreSQL
database.